The Culture Box website
This is a front end web development project with codeFirstGirls. This team is made up of Emma Shleifer, Safa Daud and Neelu Begum. The website is called The Culture Box. It provides an interactive map that shows world culture, history and information that has been deposited but people all around the world. Learn about history, literature, historical places and popular foods.